In the P6bigint body union, the flag most only be first for 64 bit Big Endian.
The intent is that the flag value (0xFFFFFFFF) is in the lower 32 bits of the
function pointer. On 64 Big Endian systems, the existing code implemented this
correctly, by storing the 32 bit value first, the 32 bit flag second. But for
32 bit Big Endian systems the flag must come first (just like all Little
Endian systems), because otherwise the 32 bit value and 32 bit pointer share
the same memory, and writing the flag doesn't cause set the lower bits of the
pointer)